# Reduce how long it takes to walk the varrefs in an expression