This is used quite a bit to determine which fields are needed in a
condition. When the condition gets large, the memory usage begins to
slow it down considerably and it doesn't take care of duplicates.

- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated