@jsternberg 2017-08-30
Benchmark comparison:

```
benchmark                old ns/op     new ns/op     delta
BenchmarkExprNames-8     131087        8389          -93.60%

benchmark                old allocs     new allocs     delta
BenchmarkExprNames-8     308            10             -96.75%

benchmark                old bytes     new bytes     delta
BenchmarkExprNames-8     135424        240           -99.82%
```